Update mixins/api for 170

+++ runescape-api/src/main/java/net/runelite/rs/api/RSDecorativeObject.java
@@ -31,7 +31,7 @@ public interface RSDecorativeObject extends DecorativeObject
 {
 	@Import("hash")
 	@Override
-	int getHash();
+	long getHash();
 
 	@Import("x")
 	int getX();
